Distort images in same folder by flipping and rotating
Import SimpleCV.  Load images in folder into a set.
For loop the images and each iteration: rotate and save
as <FILENAME>Rot.png, and flip horizontally and save as
<FILENAME>Flip.png.

This is necessary to produce more data for use in
training our Restricted Boltzmann Machines and
Convolutional Neural Network.  This process is fast.