Compute wavelet and Fourier transform of image data
The file preprocess.py in the demos directory tree loads a color image, splits it into channels, normalizes the results and computes a wavelet and Fourier transform. The resulting array of numbers can be used as features for a classifier algorithm.

Also in this commit, are three new entries into the bibliography.